Integrate Password Finder into the daemon
